pyproject.toml
[tool.poetry]
name = "fast-shum"
version = "0.1.0"
description = ""
authors = ["NIckolayEVSE  <biomaster2014@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.109.0"}
uvicorn = "^0.27.0"
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"
bcrypt = "^4.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
